get_article_text/requirements.txt
flask
newspaper3k